Renamed components folder for wing
