+++ src/fastga/models/weight/mass_breakdown/a_airframe/wing_components/compute_skin_mass.py
