+++ src/fastga/models/weight/mass_breakdown/unitary_tests/test_beechcraft_76.py
@@ -25,7 +25,7 @@ from ..a_airframe import (
     ComputeLandingGearWeight,
     ComputeWingMassAnalytical,
 )
-from ..a_airframe.components import (
+from ..a_airframe.wing_components import (
     ComputeWebMass,
     ComputeLowerFlange,
     ComputeUpperFlange,
@@ -365,9 +365,7 @@ def test_loop_compute_owe():
 
     # noinspection PyTypeChecker
     mass_computation = run_system(
-        MassBreakdown(propulsion_id=ENGINE_WRAPPER, payload_from_npax=True),
-        ivc,
-        check=True,
+        MassBreakdown(propulsion_id=ENGINE_WRAPPER, payload_from_npax=True), ivc, check=True,
     )
     oew = mass_computation.get_val("data:weight:aircraft:OWE", units="kg")
     assert oew == pytest.approx(1126, abs=1)
