+++ tests/testing_utilities.py
@@ -132,7 +132,5 @@ class Timer(object):
     def __exit__(self, type, value, traceback):
         print("\n")
         if self.name:
-            print(
-                "[%s]" % self.name,
-            )
+            print("[%s]" % self.name,)
         print("Elapsed: %s" % (time.time() - self.tstart))
